VS 100 and 110 projects
- generated VS 100 and 110 projects
- changed message length from signed 32 bit int to std::size_t

+++ MongoDB/src/RequestMessage.cpp
@@ -61,7 +61,7 @@ void RequestMessage::send(std::ostream& ostr)
 	buildRequest(requestWriter);
 	requestWriter.flush();
 
-	messageLength(ss.tellp());
+	messageLength(static_cast<std::size_t>(ss.tellp()));
 
 	BinaryWriter socketWriter(ostr, BinaryWriter::LITTLE_ENDIAN_BYTE_ORDER);
 	_header.write(socketWriter);
